+++ XMonad/Prompt/Ssh.hs
@@ -57,14 +57,14 @@ instance XPrompt Ssh where
     showXPrompt       Ssh = "SSH to: "
     commandToComplete _ c = maybe c (\(_u,h) -> h) (parseHost c)
     nextCompletion _t c l = maybe next (\(u,_h) -> u ++ "@" ++ next) hostPared
-                            where 
+                            where
                               hostPared = parseHost c
                               next = getNextCompletion (maybe c (\(_u,h) -> h) hostPared) l
 
 sshPrompt :: XPConfig -> X ()
 sshPrompt c = do
   sc <- io sshComplList
-  mkXPrompt Ssh c (mkComplFunFromList sc) ssh
+  mkXPrompt Ssh c (mkComplFunFromList c sc) ssh
 
 ssh :: String -> X ()
 ssh = runInTerm "" . ("ssh " ++ )